Fix uploader when ETag value is not quoted
Previous version of the code was blindly removing the quotes assuming
they were always present, in the case where they are not (for instance
when using it against compatible-ish backends) then the ETag was making
the request fail.

This version only removes the quotes around the value if they exist.